backend/requirements.txt
# Deep Stock Insights - Backend Dependencies
# Install with: pip install -r requirements.txt

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- Database ---
sqlalchemy==2.0.30
alembic==1.13.1

# --- Data & ML ---
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.0
lightgbm==4.5.0
tensorflow==2.16.1
keras==3.3.3

# --- Market Data ---
yfinance>=1.2.0
requests==2.32.3
aiohttp==3.9.5

# --- Technical Indicators ---
# Note: pandas-ta removed (Python 3.10 incompatibility) — indicators implemented manually

# --- Scheduling ---
apscheduler==3.10.4

# --- Utilities ---
python-dotenv==1.0.1
pydantic-settings==2.3.0
pydantic==2.7.3
email-validator==2.1.1
httpx==0.27.0
websockets>=13.0
